requirements.txt
spotipy>=2.23.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
python-dotenv>=1.0.0
jinja2>=3.1.2
pandas>=2.1.0

# Selenium for web scraping (editorial playlist fallback)
selenium>=4.15.0
undetected-chromedriver>=3.5.4
webdriver-manager>=4.0.1

# PDF generation
weasyprint>=60.0

